Small changes in compile script

+++ compile_k2.sh
@@ -7,80 +7,39 @@ then
   exit 1
 fi
 
+if [[ "$OSTYPE" == "darwin"* ]]
+then
+  echo "ERROR: Cannot compile K2 on MacOS."
+  exit 1
+fi
+
+
 # source /cvmfs/sft.cern.ch/lcg/views/LCG_101/x86_64-centos8-gcc11-opt/setup.sh
 
 cd xcoll/scattering_routines/k2/FORTRAN_src
 rm *.mod *.o
 
-
-# compile libraries
+# Compile libraries
+CFLAGS='-fPIC -O3 -mfpmath=sse -msse2 -mavx -mavx2 -mno-fma4 -mno-fma'
 cd crlibm
 make clean
 rm -r CMakeCache*.txt CMakeFiles Makefile* cmake_install*.cmake &> /dev/null
 cmake .
-make CFLAGS='-fPIC -O3 -mfpmath=sse -msse2 -mavx -mavx2 -mno-fma4 -mno-fma'
+make CFLAGS=$CFLAGS
 mv libcrlibm.a ../
 cd ../roundctl
 make clean
 rm -r CMakeCache*.txt CMakeFiles Makefile* cmake_install*.cmake &> /dev/null
 cmake .
-make CFLAGS='-fPIC -O3 -mfpmath=sse -msse2 -mavx -mavx2 -mno-fma4 -mno-fma'
+make CFLAGS=$CFLAGS
 mv libroundctl.a ../
 cd ..
 
-# compile fortran
-gfortran -m64 -fpic -funroll-loops -std=f2008 -cpp -DDOUBLE_MATH -DCRLIBM -DROUND_NEAR -O3 \
- -mfpmath=sse -msse2 -mavx -mavx2 -mno-fma4 -mno-fma -c \
- core_tools.f90 \
- constants.f90 \
- strings.f90 \
- mod_alloc.f90 \
- common_modules.f90 \
- string_tools.f90 \
- mod_units.f90 \
- extra.f90 \
- mod_particles.f90 \
- prror.f90 \
- mod_meta.f90 \
- mod_time.f90 \
- bouncy_castle.f90 \
- libcrlibm.a \
- libroundctl.a \
- coll_jawfit.f90 \
- coll_common.f90 \
- coll_db.f90 \
- mod_ranlux.f90 \
- mod_funlux.f90 \
- coll_crystal.f90 \
- coll_k2.f90 \
- coll_dist.f90 \
- collimation.f90
+# Compile FORTRAN
+files="core_tools.f90 constants.f90 strings.f90 mod_alloc.f90 common_modules.f90 string_tools.f90 mod_units.f90 extra.f90 mod_particles.f90 prror.f90 mod_meta.f90 mod_time.f90 bouncy_castle.f90 libcrlibm.a libroundctl.a coll_jawfit.f90 coll_common.f90 coll_db.f90 mod_ranlux.f90 mod_funlux.f90 coll_crystal.f90 coll_k2.f90 coll_dist.f90 collimation.f90"
+gfortran -m64 -fpic -funroll-loops -std=f2008 -cpp -DDOUBLE_MATH -DCRLIBM -DROUND_NEAR -O3 -mfpmath=sse -msse2 -mavx -mavx2 -mno-fma4 -mno-fma -c $files
 
-# link fortran
-f2py -m pyk2f -c pyk2.f90 \
- core_tools.o \
- constants.o \
- strings.o \
- mod_alloc.o \
- common_modules.o \
- string_tools.o \
- mod_units.o \
- extra.o \
- mod_particles.o \
- prror.o \
- mod_meta.o \
- mod_time.o \
- bouncy_castle.o \
- libcrlibm.a \
- libroundctl.a \
- coll_jawfit.o \
- coll_common.o \
- coll_db.o \
- mod_ranlux.o \
- mod_funlux.o \
- coll_crystal.o \
- coll_k2.o \
- coll_dist.o \
- collimation.o
+# Link FORTRAN
+f2py -m pyk2f -c pyk2.f90 ${files/.f90/.o}
 
 mv pyk2f.*.so ../
